Code + fix: fixed a weird bug where the world would be reflected, added closest point call to sphere and changed floats to doubles.

+++ includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: crypto <crypto@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/15 18:04:41 by ncarvalh          #+#    #+#             */
-/*   Updated: 2023/08/03 11:36:49 by crypto           ###   ########.fr       */
+/*   Updated: 2023/08/03 13:06:29 by crypto           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@ void	message(t_root *root, char *text);
 
 void	destroy_root(t_root **root);
 
-float	determinant(float a, float b, float c);
+double	determinant(double a, double b, double c);
 
-int	formula(float a, float b, float c, float *root);
+int	formula(double a, double b, double c, double *root);
 
-void	closest_point(float t, t_ray *ray, t_color *color);
+void	closest_point(double t, t_ray *ray, t_color *color);
 
 #endif
